refactor: Loaders.js
